Move tcp_config
