+++ nano/node/CMakeLists.txt
@@ -164,6 +164,8 @@ add_library(
   transport/message_deserializer.cpp
   transport/tcp_channels.hpp
   transport/tcp_channels.cpp
+  transport/tcp_config.hpp
+  transport/tcp_config.cpp
   transport/tcp_listener.hpp
   transport/tcp_listener.cpp
   transport/tcp_server.hpp
